Exchange balance feature:

+++ src/main/java/pl/nn/bankaccount/domain/dto/ExchangeRateDto.java
@@ -1,10 +1,11 @@
 package pl.nn.bankaccount.domain.dto;
 
+import java.math.BigDecimal;
 import pl.nn.bankaccount.domain.Currency;
 
 public record ExchangeRateDto(
         Currency currency,
-        double bid,
-        double ask
+        BigDecimal bid,
+        BigDecimal ask
 ) {
 }
