+++ src/main/java/pl/nn/bankaccount/infrastructure/currency/dto/CurrencyExchangeResponse.java
@@ -1,5 +1,6 @@
 package pl.nn.bankaccount.infrastructure.currency.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import pl.nn.bankaccount.domain.Currency;
 import pl.nn.bankaccount.domain.dto.ExchangeRateDto;
@@ -13,6 +14,6 @@ public record CurrencyExchangeResponse(
 
     public ExchangeRateDto toExchangeRateDto(Currency currency) {
         RateResponse rate = rates.getFirst();
-        return new ExchangeRateDto(currency, rate.bid(), rate.ask());
+        return new ExchangeRateDto(currency, BigDecimal.valueOf(rate.bid()), BigDecimal.valueOf(rate.ask()));
     }
 }
